django/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
Django==5.1.3
django-admin-interface==0.29.1
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.9.1
numpy==2.0.2
packaging==24.2
pillow==11.0.0
preshed==3.0.9
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
text-unidecode==1.3
thinc==8.3.2
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
django-filter==24.3
drf-yasg==1.21.8
spacy==3.8.2
python-dotenv==1.0.1
markdown==3.6
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
scikit-learn
joblib
